Make separate folder for acitivity requests, add schemas file
