+++ src/api/core/requests.ts
@@ -1,7 +1,6 @@
 import axios from 'axios'
 import { web3 } from 'fbonds-core'
 
-import { LenderActivity } from '@banx/api/core'
 import { BACKEND_BASE_URL, IS_PRIVATE_MARKETS } from '@banx/constants'
 
 import {
@@ -9,13 +8,8 @@ import {
   BorrowNftsAndOffers,
   BorrowNftsAndOffersResponse,
   BorrowNftsAndOffersSchema,
-  BorrowedActivityResponse,
-  BorrowerActivity,
-  BorrowerActivitySchema,
   FetchMarketOffersResponse,
   LendLoansAndOffersResponse,
-  LenderActivityResponse,
-  LenderActivitySchema,
   Loan,
   LoanSchema,
   Market,
@@ -306,94 +300,6 @@ export const fetchAuctionsLoans = async () => {
 
     try {
       await LoanSchema.array().parseAsync(data.data)
-    } catch (validationError) {
-      console.error('Schema validation error:', validationError)
-    }
-
-    return data.data
-  } catch (error) {
-    console.error(error)
-    return []
-  }
-}
-
-type FetchLenderActivity = (props: {
-  walletPubkey: string
-  order: string
-  sortBy: string
-  getAll?: boolean
-  skip?: number
-  limit?: number
-  collection?: string[]
-}) => Promise<LenderActivity[]>
-export const fetchLenderActivity: FetchLenderActivity = async ({
-  walletPubkey,
-  getAll = true, //TODO Remove when normal pagination added
-  order = 'desc',
-  sortBy,
-  skip = 0,
-  limit = 10,
-  collection,
-}) => {
-  try {
-    const queryParams = new URLSearchParams({
-      order,
-      skip: String(skip),
-      limit: String(limit),
-      getAll: String(getAll),
-      sortBy: String(sortBy),
-      collection: String(collection),
-      isPrivate: String(IS_PRIVATE_MARKETS),
-    })
-
-    const { data } = await axios.get<LenderActivityResponse>(
-      `${BACKEND_BASE_URL}/activity/lender/${walletPubkey}?${queryParams.toString()}`,
-    )
-
-    console.log(data, 'data')
-
-    try {
-      await LenderActivitySchema.array().parseAsync(data.data)
-    } catch (validationError) {
-      console.error('Schema validation error:', validationError)
-    }
-
-    return data.data
-  } catch (error) {
-    console.error(error)
-    return []
-  }
-}
-
-type FetchBorrowerActivity = (props: {
-  walletPubkey: string
-  order?: string
-  getAll?: boolean
-  skip?: number
-  limit?: number
-}) => Promise<BorrowerActivity[]>
-export const fetchBorrowerActivity: FetchBorrowerActivity = async ({
-  walletPubkey,
-  getAll = true, //TODO Remove when normal pagination added
-  order = 'desc',
-  skip = 0,
-  limit = 10,
-}) => {
-  try {
-    const queryParams = new URLSearchParams({
-      order,
-      skip: String(skip),
-      limit: String(limit),
-      getAll: String(getAll),
-      isPrivate: String(IS_PRIVATE_MARKETS),
-    })
-
-    const { data } = await axios.get<BorrowedActivityResponse>(
-      `${BACKEND_BASE_URL}/activity/borrower/${walletPubkey}?${queryParams.toString()}`,
-    )
-
-    try {
-      await BorrowerActivitySchema.array().parseAsync(data.data)
     } catch (validationError) {
       console.error('Schema validation error:', validationError)
     }
