+++ src/api/core/types.ts
@@ -131,7 +131,7 @@ export interface FetchMarketOffersResponse {
   meta: PaginationMeta
 }
 
-const NFTSchema = z.object({
+export const NFTSchema = z.object({
   mint: z.string(),
   meta: z.object({
     imageUrl: z.string(),
@@ -245,48 +245,3 @@ export interface AuctionsLoansResponse {
   meta: PaginationMeta
 }
 
-export const LenderActivitySchema = z.object({
-  status: z.string(),
-  publicKey: z.string(),
-  apr: z.number(),
-  duration: z.number(),
-  interest: z.number(),
-  lent: z.number(),
-  nftImageUrl: z.string(),
-  nftMint: z.string(),
-  nftName: z.string(),
-  received: z.number(),
-  timestamp: z.number(),
-  user: z.string(),
-  nft: NFTSchema,
-})
-
-export type LenderActivity = z.infer<typeof LenderActivitySchema>
-
-export interface LenderActivityResponse {
-  data: LenderActivity[]
-  meta: PaginationMeta
-}
-
-export const BorrowerActivitySchema = z.object({
-  status: z.string(),
-  publicKey: z.string(),
-  apr: z.number(),
-  borrowed: z.number(),
-  duration: z.number(),
-  interest: z.number(),
-  nftImageUrl: z.string(),
-  nftMint: z.string(),
-  nftName: z.string(),
-  repaid: z.number(),
-  timestamp: z.number(),
-  user: z.string(),
-  nft: NFTSchema,
-})
-
-export type BorrowerActivity = z.infer<typeof BorrowerActivitySchema>
-
-export interface BorrowedActivityResponse {
-  data: BorrowerActivity[]
-  meta: PaginationMeta
-}
