+++ src/pages/OffersPage/components/HistoryOffersTable/TableCells/ReceivedCell.tsx
@@ -4,7 +4,7 @@ import { BondTradeTransactionV2State } from 'fbonds-core/lib/fbond-protocol/type
 
 import { createSolValueJSX } from '@banx/components/TableComponents'
 
-import { LenderActivity } from '@banx/api/core'
+import { LenderActivity } from '@banx/api/activity'
 
 interface ReceivedCellProps {
   loan: LenderActivity
