+++ src/pages/OffersPage/components/HistoryOffersTable/TableCells/StatusCell.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import { LenderActivity } from '@banx/api/core'
+import { LenderActivity } from '@banx/api/activity'
 import { LoanStatus, STATUS_LOANS_COLOR_MAP, STATUS_LOANS_MAP } from '@banx/utils'
 
 import styles from '../HistoryOffersTable.module.less'
