+++ src/pages/OffersPage/components/HistoryOffersTable/columns.tsx
@@ -8,7 +8,7 @@ import {
   createSolValueJSX,
 } from '@banx/components/TableComponents'
 
-import { LenderActivity } from '@banx/api/core'
+import { LenderActivity } from '@banx/api/activity'
 
 import { APRCell, ReceivedCell, StatusCell } from './TableCells'
